Quick Unit test fix
