--- desktop/src/test/com/fantasticfive/projecthex/LocalGameTest.java
@@ -1,6 +1,0 @@
-package com.fantasticfive.test.projecthex;
-
-import com.fantasticfive.shared.SettingsTest;
-
-public class LocalGameTest extends SettingsTest {
-}
