--- desktop/src/test/com/fantasticfive/server/BuildingFactoryTest.java
@@ -1,6 +1,0 @@
-package com.fantasticfive.test.server;
-
-import com.fantasticfive.shared.SettingsTest;
-
-public class BuildingFactoryTest extends SettingsTest {
-}
