--- desktop/src/test/com/fantasticfive/server/RemoteGameTest.java
@@ -1,6 +1,0 @@
-package com.fantasticfive.test.server;
-
-import com.fantasticfive.shared.SettingsTest;
-
-public class RemoteGameTest extends SettingsTest {
-}
