--- desktop/src/test/com/fantasticfive/server/UnitFactoryTest.java
@@ -1,6 +1,0 @@
-package com.fantasticfive.test.server;
-
-import com.fantasticfive.shared.SettingsTest;
-
-public class UnitFactoryTest extends SettingsTest {
-}
