--- desktop/src/test/com/fantasticfive/shared/SettingsTest.java
@@ -1,42 +1,0 @@
-package com.fantasticfive.test.shared;
-
-import com.badlogic.gdx.Application;
-import com.badlogic.gdx.ApplicationListener;
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.backends.headless.HeadlessApplication;
-import com.badlogic.gdx.graphics.GL20;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.mockito.Mockito;
-
-
-public class SettingsTest {
-    // This is our "test" application
-    private static Application application;
-
-    // Before running any tests, initialize the application with the headless backend
-    @BeforeClass
-    public static void init() {
-        // Note that we don't need to implement any of the listener's methods
-        application = new HeadlessApplication(new ApplicationListener() {
-            @Override public void create() {}
-            @Override public void resize(int width, int height) {}
-            @Override public void render() {}
-            @Override public void pause() {}
-            @Override public void resume() {}
-            @Override public void dispose() {}
-        });
-
-        // Use Mockito to mock the OpenGL methods since we are running headlessly
-        Gdx.gl20 = Mockito.mock(GL20.class);
-        Gdx.gl = Gdx.gl20;
-    }
-
-    // After we are done, clean up the application
-    @AfterClass
-    public static void cleanUp() {
-        // Exit the application first
-        application.exit();
-        application = null;
-    }
-}
