--- desktop/src/test/com/fantasticfive/test/projecthex/LocalGameTest.java
@@ -1,6 +1,0 @@
-package com.fantasticfive.projecthex;
-
-import com.fantasticfive.shared.SettingsTest;
-
-public class LocalGameTest extends SettingsTest {
-}
