--- desktop/src/test/com/fantasticfive/test/server/BuildingFactoryTest.java
@@ -1,6 +1,0 @@
-package com.fantasticfive.server;
-
-import com.fantasticfive.shared.SettingsTest;
-
-public class BuildingFactoryTest extends SettingsTest {
-}
