--- desktop/src/test/com/fantasticfive/test/server/RemoteGameTest.java
@@ -1,6 +1,0 @@
-package com.fantasticfive.server;
-
-import com.fantasticfive.shared.SettingsTest;
-
-public class RemoteGameTest extends SettingsTest {
-}
