--- desktop/src/test/com/fantasticfive/test/testsettings/SettingsTest.java
@@ -1,4 +1,0 @@
-package com.fantasticfive.test.TestSettings;
-
-public class SettingsTest {
-}
